requirements.txt
Flask==3.0.3
numpy==1.26.4
Pillow==10.4.0
torch>=2.2.0
transformers>=4.44.0
torchaudio>=2.2.0
soundfile>=0.12.1
scikit-learn>=1.4.0
